requirements.txt
streamlit
langchain
langchain-community
langchain-huggingface
chromadb
python-dotenv
requests
beautifulsoup4
sentence-transformers
pypdf 